Support indexing for circular buffering (#2587)
Circular buffering is one area that the IdModel-based indexing still
doesn't work well with because the three loops generated by circular
buffering are not represented in any of ID graphs. That's why this PR
adds the loop-nest parameter to `getLinearIndex` since properties on
circular buffering are only available through actual for-loops. While
this is not ideal, I think it's more important to get a working
implementation first, so here it is.

There are mainly three changes:

- In a circular buffer loop, no matter prologue, main or epilogue, [the
loop index is not the one registered for the iter domain of the loop
only but it's also associated with the loop
stage.](https://github.com/NVIDIA/Fuser/blob/main/csrc/compute_at_map.cpp#L894-L905).
In this PR, I just added a WAR to get it from `ForLoop` if necessary
(`getLoopIndexOfCircularBufferLoop`)
- In the circular buffer main loop, reading the input tensor must be
advanced by `num_stages - 1`
(`getLoopIndexOffsetForProducerOfCircularBuffer`)
- The circular buffer tensor needs to switch buffers by adding an offset
to its linear offset, which should look like `i % num_stages`, where `i`
is the index of the circular buffer loop.
(`getOffsetForCircularBufferTensor`)