fixing output aliased tv (#2566)
Fixes: #2564 

allows nvfuser to return aliased tv output.
The updated behavior is that, when `addOutput` tries to return an
aliased output, we just need to remove the existing entry in `outputs_`,
if it exists.

The updated API involves:
```
  addOutput(Val* output, bool hide_output = false);
  aliasOutputToInput(Val* output, Val* input);
```

The expected behavior:
1. when user calls `aliasOutputToInput`, the intention is just to insert
an inplace update. Hence we are not trying to add `output` to the actual
return of the fusion.

2. when user calls `addOutput(output, hide_output)`.
2.1 if hide_output == false (default), we expect the output to be
reflected as part of the functional return. we will update
io_alias_[output].hide_output to false if it exists; we'll also remove
the existing entry of `output` in `outputs_` (since it's only used for
inplace update), and appending output to the right position.
2.2 if hide_output == true, we know that we are only adding output for
alias, so we'll just append it as-is.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>