Avoid Reduction/Broadcast domains in ValGraph scheduling
Currently, in `id_model/schedule.cpp` if a requested `ExprGroup` doesn't
already exist we select the first element in each producer `ValGroup`,
clone it without rfactor, then apply IterDomain methods `split`,
`merge`, or `swizzle`, and map the result as a new consumer `ValGroup`.

Selecting the first element of the `ValGroup` is problematic though
since it might mean we try to merge an `IterType::Reduction` domain with
an `IterType::Iteration` domain. Instead, this PR introduces a utility
`IterDomain* representativeId(const ValGroup& vg)` which picks the
"best-behaved" `IterDomain` from the group. It errors if there are no
`IterDomain`s in the `ValGroup`.

I created the utility to avoid the Iteration/Reduction merge problem,
but it is useful in other contexts also such as trying to extract a
constant extent from a `ValGroup` where there might be some mapped
symbolic extents, etc.