Clear fusion cache to fix OOM (#2619)
Fix #2617

`SchedulerRuntimeInfo` is created in `setupSchedule`. Normally,
`finalizeSchedule` is called after scheduling, which destroys
`runtime_info`. When there is an exception, the program exits. During
testing, we don't exit the program after an exception, so we accumulate
`UserSchedule` objects. The fix is to clear the `FusionCache` after each
test iteration to destroy those objects.