Print repro when error occurs at definition
We currently check for errors during `fd.execute()` and we print a
useful repro. However, some errors are caught during definition and
currently we just print the error without a repro. Instead, this PR just
copies some of the printing code to the `__exit__` method of
`FusionDefinition` so that we get a nice error message e.g.
```
from nvfuser import FusionDefinition, DataType

def nvfuser_fusion_id0(fd : FusionDefinition) -> None :
    T0 = fd.define_tensor(shape=[-1, -1, -1], contiguity=[True, True, None], dtype=DataType.Float, is_cpu=False, stride_order=[2, 1, 0])
    T1 = fd.ops.where(T0, T0, T0)

with FusionDefinition() as fd:
    nvfuser_fusion_id0(fd)

```