Support predicate indexing for unswitch/unroll/vectorize (#2640)
Adding support of generating predicates for unswitch/unroll and
vectorize.

These predicates are generated by `UnswitchPredicate`. Unswitch and
unroll are effectively the same for predication. Vectorize is also
mostly the same with only one difference, which is that the predicate of
the vectorized loop itself is altered.

For example, if a tensor has `[I0, I1, I2]` and `I1` is unswitched, the
code structure would look like:

```
for i0 in I0:
  if idx(i0, I1 - 1, I2 - 1) < N:
    for i1 in I1:
       for i2 in I2:
           do_something
       end-for
     end-for
  else
    for i1 in I1:
       for i2 in I2:
           if idx(i0, i1, i2) < N:
              do_something
           end-if
       end-for
     end-for
  end-if
end-for
```

Here, `idx` is a function to produce the linearized index using the loop
indices, and `N` is the extent of the domain. Notice that both `I1` and
`I2` loops are predicated with their maximum index values.
 
If `I1` is vectorized:

```
for i0 in I0:
  for i2 in I2:
    if idx(i0, I1 - 1, i2) < N:
      do_something
    end-if
   end-for
end-for
```

I tried to reproduce the same predicates as much as possible with the
current lowering. For example, while working on this, I realized it
should be also fine to use `0` instead of `I - 1` for vectorized domains
since splits involved in vectorized domains are guaranteed to be
divisible. For example, this should be equivalent to the above example:

```
for i0 in I0:
  for i2 in I2:
    if idx(i0, 0, i2) < N:
      do_something
    end-if
   end-for
end-for
```

Using zero would probably make the expression simplifier work more
effectively, so that's probably a better option for performance. I'll
look into it as a separate PR (#2628).