Host irs: Reduce-Scatter based pipelining (#2584)
# What

We provide a full Host IR implementation of the "Reduce-Scatter based
pipelining" technique for achieving overlap between comms and compute.
An Aten implementation has already been provided in
- #2401

This Host Ir implementation aims to reproduce the same behavior, up to
some details. In particular, the pipelining is now achieved on a single
stream, which does not allow overlap -- we'll add multi-streaming in an
upcoming PR.

### Stacked on top of:
- #2535
- #2546
- #2562

### [TODO in upcoming
PRs](https://github.com/NVIDIA/Fuser/pull/2584#issuecomment-2238833771)

# How
By the way, we added the function 
- `void ExpressionEvaluator::invalidate(Val*)` which removes values from
the cache `known_values_`. This is necessary when using
`ExpressionEvaluator` in a non-SSA context, which is the case here
because of the for-loop.
- `ir_utils:: allConsumerValsOf` which returns all (and not only direct)
consumers of vals


# NSight profile analysis and discussion

## Analysis of ATen implementation

We provided in #2401 an NSight profile analysis of the ATen
implementation we are trying to reproduce here.
- After the Nsight profile was conducted, some revisions of the PR were
applied, in particular, we switched to using `matmul_out` instead of
(unfused ops) pointwise_multiply + reduce_sum. This change makes the
compute way more efficient. Because of that, it is harder to see
overlap, because 1) the compute finishes too fast even when I'm
increasing the tensor's size to maximum capacity of the GPU, and, 2)
some new cudaMalloc appear in the profile (even for when disabling multi
stream).


- In the present PR, there is no multi-streaming so we can't hope anyway
to see any overlap. I will stick here to illustrate the relative
resemblance of the profiles obtained with Host IR and ATen
implementation, both run without multi-stream. This proves that host IR
is at least able to relatively reproduce the ATen implementation, which
is an objective in itself. Note that we see `cudaMalloc` striking again
even in the context of a single stream, both for ATen and HostIr. We
might wanna figure out where these come from before going further

<img width="2052" alt="Screenshot 2024-07-18 at 16 12 26"
src="https://github.com/user-attachments/assets/ed6ae244-77c7-447e-a7ef-2ba5df2e9d1b">

*Setup*: 8GPUs on one node: DGX V100 32GB
*Params*: LOG2_M=20, LOG2_K=6, LOG2_N=6, LOG2_S=3, MULTISTREAM=false
Left panel: ATen
Right Panel: HostIr

We can see that HostIr suffers from significant overhead from
ExpressionEvaluator And an extra `cudaMalloc`