Refactor `ValGraphBFS` to support traversing both ValGraphs and IR nodes (#2618)
Rename `ValGraphBFS` as `BFS`, move it to `bfs.h`, and refactor it as a
class template to make it support traversing both ValGraphs and IR
nodes. Now there is both a `ValGraphBFS` and an `IRBFS`. They are
subclass of `BFS` with different template arguments. `ValGraphBFS` is
used to traverse ValGraph, and `IRBFS` is used to traverse IR nodes.

This PR also use `IRBFS` to rewrite the current `ir_utils::allIDsOf` to
make sure all IDs are obtained when domains are not on a linear path.
Also, it is used to modify `validateDomainEquivalence` to remove the
restrictions the previous implementation has.