Promote hoisting of vectorize predicates (#2628)
Currently, we generate predicates for vectorization as:

```
// 4-way vectorization
if (idx(i + 3) < N) {
  local_tensor[0] = loadGlobalToLocal(gmem_tensor[idx(i)]);
}
```

where `idx(x)` returns an index of the tensor with the vectorized loop
index as `x`. The predicate is to make sure that the access of the 4-way
vectorized load is within the extent.

This should be equivalent to:

```
// 4-way vectorization
if (idx(i) < N) {
  local_tensor[0] = loadGlobalToLocal(gmem_tensor[idx(i)]);
}
```

Notice the predicate is the same as the index for the load access. I
belive that these two patterns should be equivalent since we assert that
the vectorized domain is divisible by the vectorization factor.

The latter pattern is likely to promote more effective hoisting as the
indices of the predicate and the tensor access are the same. Also, I
think we can simplify the predicate logic for vectorization.

I tested the change with all the C++ tests and the python benchmarks
using compute-sanitizer on H100. No invalid memory access happened,
although sanitizer may not be perfect and also some of the matmul tests
couldn't run due to out-of-memory errors (likely because sanitizer needs
more memory).

Anyone has any concern?