Fix incorrect DataType in allocGlobalBufferForGridComm (#2636)
In #2634, an `IterDomain` is created during lowering (in
`allocGlobalBufferForGridComm`) using an extent that is derived from
parallel extents. In this case the computation is done with `Int`
scalars, and the constructor of `IterDomain` asserts that the extent
should be `Index` instead. This PR casts the stop value to `Index` if
necessary before creating the `IterDomain` in
`allocGlobalBufferForGridComm`.

The reason we have an `Int` parallel dimension is that we compute
`SimplifyExpr::maxExpr` with two constant `IterDomain` extents each of
which is `DataType::Index` and should return `DataType::Index`. However,
it currently returns `Int`. This is similar to the problem encountered
in #2646. This PR fixes this by computing the proper output type for
`maxExpr`, `minExpr`, `bitwiseAndExpr`, and `bitwiseOrExpr`.

Fixes #2634