Allow creating broadcast from nowhere (#2459)
Make it possible to create a new broadcast during scheduling. This new
broadcast neither comes from a `BroadcastOp` nor from a fusion input. It
just appears from no where. This PR only adds the minimal support to our
system to make the simple pointwise fusion at test
`BroadcastFromNowhere` work. So far, we can not inline such fusions. Our
sync analysis could not handle this either. I am not working on these
two pieces so far, because it is not clear to me whether this will be
needed for scheduling hopper MMA.

---------

Co-authored-by: Jacob Hinkle <1454944+jacobhinkle@users.noreply.github.com>