Make sure unswitch predicates propagate maximum (or minimum) indices through merge-inner or split-outer domains (#2689)
(Stacked on top of #2677)

The original issue is #681. It was addressed in #687.

This PR is NOT as comprehensive as #687, but my gut feeling is that this
should be good enough, in particular since contig indexing would avoid
backward traversals through merge in many cases.

I'll do final more comprehensive comparison with the legacy indexing
once contig indexing is done.

Since the original PR and issue were reviewed by @zasdfgbnm, could you
please review this too?