Preserve definition of IDs mutated because of mutated extents
In #2671, the replaceSymbolicSizes lowering pass calls
ir_utils::replaceValue with a seemingly benign list of scalar Val
replacements. However, an error is encountered because in replacing
IterDomains whose extents should be replaced, we wind up erasing the
definition of a Split output. We should instead preserve these
definitions and just replace the output of that expression.

Fixes #2671