Some refactor of matmul schedule (#2709)
- Remove `TensorView::is_mma_swizzled_`. This flag was designed to
prevent further schedule of loop domain, which I don't think makes sense
any more because we no longer have a root->loop linear transformation
history.
- Rename `WarpMmaSwizzler` as `MmaSwizzler`. Mma swizzling is no longer
limited to a warp, but can also be a warp group.
- Rename `scheduleMmaWarpOutput` as `scheduleMmaOutputAllocation`. The
thing that this function schedules is the allocation domain. The loop
domain may be the same as allocation domain, or may not (for the planned
hopper).
- Make `scheduleMmaOutputAllocation` accept an `AbstractTensor` and
return an `AbstractTensor`.
- Stop using `applyMmaSwizzle` to schedule matmul output. We should
prefer using `scheduleMmaOutputAllocation` directly.