Host irs: add multistreaming to rs based pipelining (#2621)
# What

Adds multi-streaming in the host IR implementation of Reduce-scatter
based pipelining that is introduced in
- #2584

It addresses item 4 of this [TODO
list](https://github.com/NVIDIA/Fuser/pull/2584#issuecomment-2238833771)

Stacked on top of:
- #2665
# Why

Multi-streaming is central for achieving overlap between comms and
compute

# How

We need to change stream at each iteration of the for-loop. To do so, we
need a way to refer to the `Stream` using the dynamic value of the
for-loop's index. Since this dynamic value is only available at runtime,
we make a `Stream` point to a `Val`, which gonna be "expression
evaluated" at runtime. This is the only way I can figure out how to
refer to a stream dynamically.



### Stacked on top of
- #2584