[codediff] Gracefully handle starting from detached HEAD (#2740)
Currently if you are on a detached HEAD (i.e. if you have run `git
checkout <SHA>` then `compare_codegen.sh` will fail with `fatal: ref
HEAD is not a symbolic ref`. This PR makes it safe to run in such a
scenario.

NOTE: I am sometimes seeing new git branches introduced at cleanup named
like the SHA of the commit and the compare commit. This has happened
previously and I think it is unrelated to this change. Since it only
affects cleanup, I would like to merge this PR so that we can improve
`!build --diff` in the CI and look into the branch issue later.