Remove `query/key_seq_len, debug_attn_mask` in `SdpaFwdOp` (#2743)
1. Since, we have deviated from torch API, making the following changes
to have a leaner API:

- Removes `debug_attn_mask` which is a 1D zero-tensor since
`return_debug_mask=false` in `SdpaFwdOp`.
- Removes `query_seq_len, key_seq_len`: These variables can be computed
in the backward pass for non-nested tensors.

2. Adds `DIDx` axis to `log_sumexp` in `SdpaFwdOp::evaluate` and
`logical_domain_mapping`. The `DIDx` axis is assumed when the tensorview
is created but was not added everywhere. This removes discrepancy in
different parts of the code. This will be revisited after work on
sharded `SdpaBwdOp`. The `DIDx` can be removed from `log_sumexp` if not
required.