Add serialization check to the pytest framework. (#2753)
This PR adds a serialization check to the pytest opinfo framework.

Details:
- The pytest framework `serde_check` only checks per-operator and
per-dtype. It doesn't check every single sample input. We check more
input variations in the pytest benchmarks than
`test_python_frontend.py`. This is a timesaving measure.
- Creates `atexit_serde_create_op_test` pseudo decorator to test
automatic serialization. Serializing error test cases embeds errors in
the serialized binary. Normally, we call `FusionCache.reset()` to clear
the cache after running an error test. In the pytest framework, the
error tests are separate from the correctness tests. We use this
decorator to apply this to the correctness tests.
- The pytest framework and `test_python_frontend.py` use different
arguments. `basic_serde_check` is the common part between them. It
serializes the cache, deletes it, and then deserialized to recreate the
cache.