Sharded SDPA backwards support (#2826)
Adds temporary support for sharded backwards scaled dot product
attention. Until https://github.com/NVIDIA/Fuser/issues/2563 is
completed.

Similar to https://github.com/NVIDIA/Fuser/pull/2565
Similar restrictions:
1. All necessary sharded inputs are manually sharded before the
SDPABwdOp is created. We cannot rely on sharding propagation or sharding
after the Fusion is created, because the dimension checks are called
when the op is created.
2. Only the head dimension is sharded and all inputs and outputs have
either a sharded head dimension or unshaded.
3. DID axis is the outermost axis. This is because during evaluation if
we see 5 dimensions, it is assumed the first is the DID axis and is
appropriately squeezed from the inputs and unsqueezed onto the outputs.