Support contiguous indexing (#2752)
This PR enables contiguous indexing in the new indexer. It's only about
indexing of tensor expressions. Predicates will be addressed as a
follow-up PR.

The basic idea is when there's a merge of two contiguous allocation
domains, we can use the merge output domain to index the 2D space
represented by the two allocation domains. This analysis exists for the
existing indexer, mostly in csrc/contiguity.cpp. This PR attempts to
reuse it as much as possible with minimal changes.

The main difference between the existing and the new indexing methods is
how the index propagation is done. In the existing method, it's always
done through the iter domains of a tensor, whereas in the new method
it's done over the AlmostExact graph. Accordingly, the contiguity
analysis is only done from the root to loop domains of a tensor, but
with the new indexer, it also needs to be done through the graph
traversal path. That's the main adaptation done in this PR.

In order to reuse the existing analysis, I created a new class
`OrderedIdGroupInformation`, which is a subclass of
`OrderedIdInformation`. For `ContigIDs`, I just created a new class
`ContigIDGroups`, which is not a subclass of `ContigIDs`, but still
pretty similar to `ContigIDs`. Once everything is moved to using the new
indexer, we could clean up the existing analysis, but for now keeping
these two classes seems to be necessary.

Follow-up TODOs
- Contig indexing for predicate 
- Contig analysis for both forward and backward traversal