Introduce multi matmul scheduler (#2719)
This is the first in a series of PRs that represents a refactor of the
matmul scheduler to increase the flexibility in the number of operands
and matmul patterns we can schedule in a single kernel. This PR:
- Introduces the `MultipleMatmulScheduler` class which will eventually
implement and extend the functionality in the current `scheduleMatmul`.
- Introduces a new option `EnableOption::FuseMultipleMatmuls` which
enables this scheduler in place of the default matmul scheduler.
- Implements some basic functionality in `MultipleMatmulScheduler`.
Enough to translate matmul patterns and define most of the relevant
tensors (`splitk_sum` and `smem_epilogue` will be defined and all
tensors will be scheduled in later PRs).

In the next PRs I will
- Schedule the prologue and introduce parametrized tests to check that
the scheduled fusion matches the current scheduler
- Schedule epilogue tensors without split-K and smem_epilogue
- Schedule split-K sum
- Schedule smem_epilogue
- Update the heuristic to adjust parameters when multiple matmul
patterns are detected. This is needed to ensure we don't use
register/smem buffers that are too large.
- Update canSchedule checks to accept multiple matmuls when the multi
matmul scheduler is enabled