Use loop promotion and indexing traversal graph to find mismatched parallelization (#2875)
See #2850
Stacked on #2901 

The old code is still used by default. With `NVFUSER_ENABLE=id_model`,
the new analysis is used. It's also used for tensors with
non-conventional domains.

This is required for #2851. It also enables previously disabled
parallelization of the mismatching reshape test from #2684.

I validated the change by comparing the results between the existing and
new analyses with all the tests and benchmarks. The only mismatch was
with the mismatching reshape test, for which the existing analysis
declared a sync is required, whereas the new one correctly recognizes
there's no cross-thread dependency.