Scheduling tool: representativeId(const ValGroup&)
This introduces a tool to get a representative IterDomain for a
ValGroup. A ValGroup might have IterDomains with different IterTypes.
For example, the Exact graph maps Iteration domains with Reduction
domains, and the permissive graph maps Broadcast domains with Iteration
when the broadcasts are concretized.

The purpose of `representativeId` is to find an `IterDomain` that is
most useful for scheduling. In particular, it is an ID that is easiest
to merge with other IDs. Since we restrict merging Reduction domains
with non-Reduction domains, we give the lowest priority to Reduction
domains in the ValGroup; we only return a Reduction domain if there is
no other choice. Other than that, we prefer Iteration to Broadcast, we
prefer unexpanded Broadcasts to expanded ones, and we prefer IDs with
constant extents to those with symbolic extents.