Contiguous predicate indexing (#2860)
This is a follow-up to #2752 and concludes contiguous indexing support
in the new indexing system. The previous only supported tensor indexing.
This one extends that for predicate indexing. The primary differences
with the tensor indexing case are:

- In predicate indexing, all logical domains can be considered
contiguous. Ordering of merge doesn't matter, either. This is also the
same with the current indexer
- However, contiguity analysis needs to take non-divisible splits into
consideration as it's invalid to do contiguous indexing through
non-divisible splits. This is also the same with the current indexer.