Concretize trivial Resize
This changes concretization info to not just hold the IterType of
dynamic Resized IterDomains but also whether the Resize is trivial.
Trivial resizes occur because both the left and right expand value is
zero. We evaluate these at concretization to determine if the Resize is
trivial. If so, we replace the Resized ID with the unresized root ID in
the logical domain.