Enable IdModel-based predication for new producer-based scheduling tests (#2935)
Fixes #2924 

Tensor indexing automatically uses the new indexer, but that wasn't the
case for predicate indexing, and because of that no proper predicate was
generated and resulted in the sanitizer errors.

Explicitly opted in using the new indexer in those tests. Also, make it
more automatic whenever non-convention scheduling is used.