Allow loop domains to have extra IDs (#2937)
Relaxes the validation of `setLoopDomain`. All IDs of the logical
domain, except for broadcast, must be reachable from the loop domain,
but the opposite doesn't need to be true.

The main use case is to set a concrete ID for a broadcast logical ID
when inlined into tensors with concrete IDs only.

```
// [i0, i1]
  auto tv0 = makeSymbolicTensor(2);
  fusion.addInput(tv0);
  // [i0]
  auto tv1 = makeSymbolicTensor(1);
  fusion.addInput(tv1);

  // [i0]
  auto tv2 = set(tv1);
  // [i0, b1]
  auto tv3 = broadcast(tv2, {false, true});
  // [i0, i1]
  auto tv4 = add(tv0, tv3);
  fusion.addOutput(tv4);
```

In this fusion, when inlining tv2 and tv3 to tv4, it effectively have
tv2 and tv3 get the same loop domain as `tv4`. That can be explicitly
done by something like this (almost):

```
 // Set the loop domain of tv2 the same as tv4. The new loop domain
  // includes an ID that is not reachable from tv2 logical domain
  tv2->setLoopDomain(
      {tv2->getLogicalDomain().at(0),
       tv4->getLogicalDomain().at(1)->cloneWithoutRFactor()});

  // Same for tv3
  tv3->setLoopDomain(
      {tv3->getLogicalDomain().at(0),
       tv4->getLogicalDomain().at(1)->cloneWithoutRFactor()});
```

This is not complete yet as the clone of tv4 logical ID is not mapped
with the tv4 logical ID itself, but assuming that's somehow done,
inlining would be trivial as all tensors use the mapped loop domains.

Note that modifying the loop domain like this way is still considered
experimental, and in a long term we would probably need a validation
that is a little more restrictive.

Not strictly required but related to #2851 and #2902.