Introduce TaggedAbstractTensor (#2933)
This generalizes `AbstractTensor` to the templated struct
`AbstractTensorWithInfo<Info>` and introduces a special case subclass
`TaggedAbstractTensor<Tag>`. This can be used by passing an enum class
for `Tag`, and holds an `unordered_set<Tag>` for each dimension. Merging
and swizzling unions these sets, and split duplicates the set.

Note that a lot of code had to be moved out of the cpp into the header
because of templatization. However, there are no changes to the
`Dispatch*` classes. The `AbstractTensorWithInfo` methods like split,
merge, swizzle, etc. are just changed to add calls to `Info::merge`.

Related to #2913, which specializes this as `using AbstractMatmulTensor
= TaggedAbstractTensor<MatmulDimRole>`.