Append to AbstractTensor's info whenever we append to domain
This fixes failures in `TMAIndexingTest` tests that result in a segfault
due to unchecked accesses that result when `domain` has entries but
`info` is empty in `AbstractTensorWithInfo`. Note that another approach
would be to switch to a class instead of a struct and add accessors that
guarantee we maintain the same length for each of these vectors.

This change also adds more assertions in scheduling methods of
AbstractTensorWithInfo to check that the `domain` and `info` attributes
maintain the same size.