Convert AbstractTensor from struct to class
`AbstractTensor` now has many methods and as of #2933 it is subclassed
into `TaggedAbstractTensor` which has additional methods. Also, there is
no longer just a single vector attribute `domain` but also `info`.
Instead of requiring the user to manage those vectors and keep them the
same length, instead this PR makes them `protected` and uses the
existing accessors to adjust them, as suggested in
https://github.com/NVIDIA/Fuser/pull/2963#discussion_r1765951724.