Fix finding normalization resolution points by traversing both forward and backward directions (#2946)
Finding resolution points fails with a persistent pattern that has
non-straight line dependencies. For example, here's a simplified pattern
from the repro of #1123:


![Scratch-31](https://github.com/user-attachments/assets/cac62d10-acee-4f4b-89e2-71c19d7a5557)

`T3` is the persistent tensor in this case, and `T7` is the resolution
point. While may not be apparent immediately, `T3` cannot be inlined
into `T4` or `T9` since that would require `T2` and `T7` and also `T6`
and `T5` to be inlined as well, but because of the reduction path, that
is not possible.

The existing resolution analysis only looks at consumers of the
persistent buffer, so it first looks at `T9` but it stops there and
fails to detect the resolution point.

This PR adds a more thorough analysis that traverses both forward and
backward directions. In the above case, after visiting `T9`, it moves on
to `T2` and then find the resolution point at `T3`.

I originally planned to completely replace the existing analysis with
the new one, but looks like it opened a can of worms, so I backed off
from the plan and only used the new analysis as a fallback analysis of
the existing one. Any fusion that works with the existing analysis
should not be affected as the fallback path should not be taken. Only
those that would fail with the existing analysis should use the new
analysis. A simplified C++ repro extracted from #1123 is also added.

Regarding the issues in replacing the existing analysis, there are some
small number of test cases where the old and new analyses result in
generating different resolution points. Some of them seem to be a matter
of what exaction resolution should mean. Since it isn't urgent, I
decided not to pursue further. One interesting case, however, was a
fusion with a persistent buffer that doesn't actually seem to be
necessary to be persistent (see #2954). I'm not sure how important that
would be, but I'll leave it as an open issue as well.

Closes #1123