Add an experimental option to map cloned IterDomains in the Exact graph (#2977)
The primary use case is when setting a loop domain of a tensor that
would be involved in broadcast forwarding in the conventional inlining
approach. For example, think of a fusion like below:

```
t0 = [i0]
t1 = [i0, i1]
t2 = broadcast(t0, {false, true}); // [i0, b2]
t3 = t1 + t2
```

To inline all tensors together, t0 and t2 would be inlined with
broadcast forwarding.

Suppose a loop domain of t3 would be set for all tensors, t0, for
example, would have a loop domain of: {t0->axis(0), clone of
t3->axis(1)}. In this case, however, the clone is almost an exact copy
of the t3 ID, there's no use or def to define its exact mapping with the
original t3 ID. This is problematic when inlining t0 since the inlining
rule would use the Exact (or Broadcast) graph to determine the validity
of inlining, but for the cloned ID of t0, since there's no def or use,
it would have no mapping with any of the other IDs, which means the
inlining would fail.

The mapping option added in this PR workarounds the issue. While I am
not completely confident about forcing certain ID mappings, I think it's
a reasonable compromise between safety and flexibility. Besides, the
mapping is currently only allowed if two IDs are the same per
IterDomain::sameAs.